Jamie uploading completed delete md

+++ php/delete-user-functions.php
@@ -6,13 +6,19 @@
 		if(!$mysqli->query("USE researchdatabase")){
 			die("Failed to use database");
 		}
+		//If a supervisor need to replace first before finishing deletion
+		$query = "SELECT name FROM study WHERE supervisor_id = $userID";
+		$result = $mysqli->query($query);
+		if(!$result){
+			die("Bad Query 2");
+		}
+		if($result->num_rows >0){
+			header('Location: /user/delete-md.php?userID='.$userID);
+			return;
+		}
 		$query = "DELETE FROM view_edit WHERE user_id = $userID";
 		if(!$mysqli->query($query)){
 			die("Bad query");
-		}
-		$query = "UPDATE study SET supervisor_id=NULL WHERE supervisor_id = $userID";
-		if(!$mysqli->query($query)){
-			die("Bad query2");
 		}
 		$query = "DELETE FROM works_on WHERE assistant_id = $userID";
 		if(!$mysqli->query($query)){
@@ -31,6 +37,34 @@
 			die("Bad query5");
 		}
 		header('Location: /users.php?successfulDeleteUser');
+	}
+	function deleteMD(){
+		
+		if(!isset($_GET['deleteMD'])){
+			return;
+		}
+		$userID = $_GET['userID'];
+		$mysqli = new mysqli("localhost", "root", "", "researchdatabase");
+		if(!$mysqli->query("USE researchdatabase")){
+			die("Failed to use database");
+		}
+		$newSupervisorName = $_POST['supervisor'];
+		$query = "SELECT id FROM person WHERE name = '$newSupervisorName'";
+		$result = $mysqli->query($query);
+		if(!$result){
+			echo $query;
+			die("Bad query2");
+		}
+		$data = $result->fetch_assoc();
+		$newSupervisorID = $data['id'];
+		$query = "UPDATE study SET supervisor_id = $newSupervisorID WHERE supervisor_id = $userID";
+		if(!$mysqli->query($query)){
+			echo $query;
+			die("bad Query 1");
+		}else{
+			deleteUser();
+		}
+	
 	}
 
 ?>
